custom_pc_odoo_17_v6/tests/test_product.py
```python
from odoo.tests import common

class TestProduct(common.TransactionCase):

    def setUp(self):
        super(TestProduct, self).setUp()
        self.Product = self.env['product.product']
        self.product_1 = self.Product.create({'name': 'Test Product 1'})

    def test_get_product_details(self):
        product_details = self.product_1.get_product_details()
        self.assertEqual(product_details['name'], 'Test Product 1')

    def test_inventory_integration(self):
        self.product_1.stock_level = 10
        self.product_1.reorder_level = 5
        self.product_1.check_stock_levels()
        self.assertEqual(self.product_1.reorder_alert, False)

        self.product_1.stock_level = 4
        self.product_1.check_stock_levels()
        self.assertEqual(self.product_1.reorder_alert, True)

    def test_product_categorization(self):
        self.product_1.category = 'CPU'
        self.assertEqual(self.product_1.category, 'CPU')

    def test_product_images(self):
        self.product_1.image = 'test_image.jpg'
        self.assertEqual(self.product_1.image, 'test_image.jpg')

    def test_product_specifications(self):
        self.product_1.specifications = 'Test specifications'
        self.assertEqual(self.product_1.specifications, 'Test specifications')
```